requirements.txt
tensorboard
easydict
mxnet
onnx
sklearn
opencv-python
torch
Pillow
retinaface
numpy